Move gazebo and launch file cli functions to generators
